refactor(structure): move source files to fitting locations
